backend/requirements.txt
flask==3.0.3
requests==2.31.0
groq==0.9.0
ics==0.7.2
python-dotenv==1.0.1
